+++ doc/version_history.rst
@@ -14,6 +14,7 @@ v1.12.0
 * Introduce a new simulation mode, where the MockController doesn't get started by the CSC, for test purposes.
 * Improve starting and stopping of MockController.
 * Improve error handling of the CSC 'write then read' loop.
+* The mock AMCS now uses the true current consumption.
 
 Requires: